requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
hl7apy==1.3.4
fhir.resources==7.1.0
jinja2==3.1.3
python-multipart==0.0.9
pydantic==2.6.4
fpdf2==2.8.1
